move network containers into their own file. Change selector arg to be a union
